requirements.txt
aiohttp==3.9.3
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
cohere==5.15.0
colorama==0.4.6
cryptography==44.0.2
Deprecated==1.2.18
devtools @ file:///home/conda/feedstock_root/build_artifacts/python-devtools_1735148275205/work
distro==1.9.0
dotenv==0.9.9
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1745502089858/work
fastavro==1.10.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
jiter==0.9.0
multidict==6.4.3
numpy==2.2.5
openai==1.77.0
packaging==25.0
pandas==2.2.3
propcache==0.3.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyGithub==2.6.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
PyJWT==2.10.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sniffio==1.3.1
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.0
